old stuff oved to old_stuff_csv
